scavanges some old electron related code
This takes some ancient code for Grist + electron and dusts it
off a bit. Doesn't do anything for packaging, that part is hard
to extract from cruft it is entangled with.

On Linux at least, this code gets as far as this sequence running:
  yarn run build:prod
  yarn run install:python
  yarn run electron
which runs Grist in electron (without packaging)